% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl.density.plot.R
\name{cyl.density.plot}
\alias{cyl.density.plot}
\title{cyl.density.plot}
\usage{
cyl.density.plot(x, fx, axis = TRUE)
}
\arguments{
\item{x}{Matrix with three columns containing the points where the density is evaluated in cartesian coordinates.}

\item{fx}{Vector containing the values of the density. The length must coincide with the number of rows of x.}

\item{axis}{Logical; if TRUE, the axis are plotted.}
}
\description{
Function \code{cyl.density.plot} creates an interactive 3D representation of a given cylindrical density.
}
\examples{
N<-600
u = seq(0, 2 * pi, length.out = N)
v = seq(0,1, length.out = N)
ll<-expand.grid(u,v)
u2<-ll[,1]
v2<-ll[,2]
T<-diff(range(v))/4
x<-cbind(v2,  T * cos(u2), T * sin(u2))
fx<-dirdensity:::dms(u2, v2, mu=c(pi,0), kappa=5, sigma=0.5, nu=0, lambda=1)
cyl.density.plot(x,fx)
}
