% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cylinder.plot.R
\name{cyl.plot}
\alias{cyl.plot}
\title{cyl.plot}
\usage{
cyl.plot(
  x,
  col.points = "red",
  pch = 16,
  size = 4,
  axis = TRUE,
  zoom = 0.9,
  userMatrix = NULL,
  windowRect = NULL,
  range = NULL,
  T = NULL,
  col.cyl = "honeydew1",
  labelsx = NULL,
  labelsy = NULL
)
}
\arguments{
\item{x}{Matrix with two columns containing the data in cylindrical coordinates, where the number of rows is the number of observations. The first column contains the azimuthal angle (in radians), while the second column contains the height.}

\item{col.points}{Vector containing the mean direction of the von Mises-Fisher guide.}

\item{pch}{Type of points to be plotted.}

\item{size}{Numerical value containing the concentration of the von Mises-Fisher guide.}

\item{axis}{Logical; if TRUE, axis are plotted.}

\item{zoom}{Parameter controlling the zoom in the plot.}

\item{userMatrix}{A 4 by 4 matrix describing user actions to display the scene. See par3d.}

\item{windowRect}{A vector of four values indicating the left, top, right and bottom of the displayed window (in pixels). Applies to the whole device. See par3d.}

\item{range}{Vector with two components giving the interval of the real-line where height of the cylinder should be displayed.}

\item{T}{If plot=TRUE, parameter controlling the radius of the cylinder.}

\item{col.cyl}{Character string indicating the desired color of the cylinder.}

\item{labelsx}{Numeric vector containing the values to be displayed as labels for the circular component.}

\item{labelsy}{Numeric vector containing the values to be displayed as labels for the linear component.}
}
\description{
Function \code{cyl.plot} creates an interactive 3D representation of cylindrical data on the surface of a cylinder.
}
\examples{
library(circular)
a1<-rvonmises(150,0,5)
a2<-runif(150,2,10)
x<-cbind(a1,a2)
cyl.plot(x)
}
