% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl.density.models.R
\name{dAL.cyl}
\alias{dAL.cyl}
\title{dAL.cyl}
\usage{
dAL.cyl(
  mu,
  lambda,
  beta,
  alpha,
  kappa,
  rangex,
  T = NULL,
  plot = TRUE,
  axis = TRUE,
  orientation = "horiz"
)
}
\arguments{
\item{mu}{Circular location parameter.}

\item{lambda}{Skewness parameter (must belong to (-1,1)).}

\item{beta}{Linear dispersion (has to be positive).}

\item{alpha}{Shape parameter (has to be positive).}

\item{kappa}{Circular concentration, which acts also as dependence parameter (greater or equal than zero).}

\item{rangex}{Vector with two components giving the interval of the real-line where linear part of the density should be evaluated.}

\item{T}{If plot=TRUE, parameter controlling the radius of the cylinder.}

\item{plot}{Logical; if TRUE, the 3D plot is produced.}

\item{axis}{Logical; if TRUE, axis are plotted.}

\item{orientation}{If plot=TRUE, character string indicating the orientation of the cylinder. Must be "horiz" or "vert".}
}
\value{
An list containing the following components: \item{fx}{ The estimated values of the density.}
\item{eval.points}{The points where the estimated density was evaluated.}
}
\description{
Function \code{dAL.cyl} computes the density function of a Abe-Ley density and produces an interactive 3D representation of the density on the surface of a cylinder.
}
\examples{
al<-dAL.cyl(mu=0,lambda=-0.5,kappa=1,beta=2,alpha=3,orientation="vert",rangex=c(0,3),T=1.5)
}
\references{
Abe, T. and Ley, C. (2017) 'A tractable, parsimonious and flexible model for cylindrical data, with applications', Econometrics and Statistics, 4, 91-104.
}
