% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.density.models.R
\name{dcos.torus}
\alias{dcos.torus}
\title{dcos.torus}
\usage{
dcos.torus(mu, kappa, rho, plot = TRUE, axis = TRUE)
}
\arguments{
\item{mu}{Vector with two components representing the mean direction of each marginal model.}

\item{kappa}{Vector with two components representing the concentration of each marginal model.}

\item{rho}{Numeric element containing the correlation between the two components.}

\item{plot}{Logical; if TRUE, the 3D plot is produced.}

\item{axis}{Logical; if TRUE, axis are plotted.}
}
\value{
An list containing the following components: \item{fx}{ The estimated values of the density.}
\item{eval.points}{The points where the estimated density was evaluated.}
}
\description{
Function \code{dcos.torus} computes the density function of a bivariate cosine model and produces an interactive 3D representation of the density on the surface of a cylinder.
}
\examples{
co<-dcos.torus(mu=c(-pi/16,pi/4), kappa=c(4,5), rho=3)
}
\references{
Mardia, K. V., Taylor, C. C. & Subramaniam, G. K. (2007), ‘Protein bioinformatics and mixtures of bivariate von Mises distributions for angular data’, Biometrics 63, 505–512.
}
