% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl.density.models.R
\name{dJW.cyl}
\alias{dJW.cyl}
\title{dJW.cyl}
\usage{
dJW.cyl(
  mu,
  lambda,
  kappa,
  rangex,
  T = NULL,
  plot = TRUE,
  axis = TRUE,
  orientation = "horiz"
)
}
\arguments{
\item{mu}{Circular location parameter.}

\item{lambda}{Linear dispersion (positive).}

\item{kappa}{Dependence parameter (has to be less than lambda).}

\item{rangex}{Vector with two components giving the interval of the real-line where linear part of the density should be evaluated.}

\item{T}{If plot=TRUE, parameter controlling the radius of the cylinder.}

\item{plot}{Logical; if TRUE, the 3D plot is produced.}

\item{axis}{Logical; if TRUE, axis are plotted.}

\item{orientation}{If plot=TRUE, character string indicating the orientation of the cylinder. Must be "horiz" or "vert".}
}
\value{
An list containing the following components: \item{fx}{ The estimated values of the density.}
\item{eval.points}{The points where the estimated density was evaluated.}
}
\description{
Function \code{dJW.cyl} computes the density function of a Jones-Wehrly density and produces an interactive 3D representation of the density on the surface of a cylinder.
}
\examples{
jw<-dJW.cyl(mu=pi/2,lambda=2,kappa=1.25,orientation="horiz",rangex=c(0,5))
}
\references{
Johnson, R. A. & Wehrly, T. E. (1978), ‘Some angular-linear distributions and related regression models’, J. Amer. Statist. Assoc. 73, 602–606.
}
