% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.density.models.R
\name{dkent.sph}
\alias{dkent.sph}
\title{dkent.sph}
\usage{
dkent.sph(mu, kappa, beta, G, plot = TRUE, axis = TRUE)
}
\arguments{
\item{mu}{Vector with three components giving the mean of the density in cartesian coordinates.}

\item{kappa}{Parameter controlling the concentration of the density.}

\item{beta}{Parameter controlling the ovalness of the density.}

\item{G}{A 3 x 3 matrix with first column equal to the mean direction. The second and third columns are the major and minor axes respectively.}

\item{plot}{Logical; if TRUE, the 3D plot is produced.}

\item{axis}{Logical; if TRUE, axis are plotted.}
}
\value{
An list containing the following components: \item{fx}{ The estimated values of the density.}
\item{eval.points}{The points where the estimated density was evaluated.}
}
\description{
Function \code{dkent.sph} computes the density function of a Kent density of dimension 2 and produces an interactive 3D representation of the density on the surface of a sphere.
}
\examples{
ken<-dkent.sph(mu=c(-1,0,0),kappa=5,beta=3,G=cbind(c(-1,0,0),c(0,1,0),c(0,0,1)))
}
\references{
Kent John (1982). The Fisher-Bingham distribution on the sphere. Journal of the Royal Statistical Society, Series B, 44(1): 71–80.
Mardia, K. V. & Jupp, P. E. (2000), Directional Statistics, Wiley, New York.
}
