% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.density.models.R
\name{dmixvMF.sph}
\alias{dmixvMF.sph}
\title{dmixvMF.sph}
\usage{
dmixvMF.sph(mu, kappa, probs, plot = TRUE, axis = TRUE)
}
\arguments{
\item{mu}{Matrix with three columns and number of rows equal to the number of components in the mixture, giving the mean vectors for all components.}

\item{kappa}{Vector of parameters controlling the concentration of each component.}

\item{probs}{Vector of parameters controlling the mixing probabilities.}

\item{plot}{Logical; if TRUE, the 3D plot is produced.}

\item{axis}{Logical; if TRUE, axis are plotted.}
}
\value{
An list containing the following components: \item{fx}{ The estimated values of the density.}
\item{eval.points}{The points where the estimated density was evaluated.}
}
\description{
Function \code{dmixvMF.sph} computes the density function of a mixture of von Mises-Fisher densities, of dimension 2, and produces an interactive 3D representation of the density on the surface of a sphere.
}
\examples{
mixvm<-dmixvMF.sph(mu=matrix(c(0,0,1,0,0,-1),ncol=3,byrow=TRUE),kappa=c(5,5),probs=c(0.6,0.4))
}
\references{
Kurt Hornik and Bettina Grun (2014). movMF: An R Package for Fitting Mixtures of von Mises-Fisher Distributions http://cran.r-project.org/web/packages/movMF/vignettes/movMF.pdf
}
