% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl.density.models.R
\name{dMS.cyl}
\alias{dMS.cyl}
\title{dMS.cyl}
\usage{
dMS.cyl(
  mu,
  kappa,
  sigma,
  nu,
  lambda,
  rangex = NULL,
  T = NULL,
  plot = TRUE,
  axis = TRUE,
  orientation = "horiz"
)
}
\arguments{
\item{mu}{Vector with two components: the first one is the circular mean and the second one is the linear mean.}

\item{kappa}{Concentration parameter (greater or equal than zero).}

\item{sigma}{Standard deviation (has to be positive).}

\item{nu}{Parameter between -pi and pi.}

\item{lambda}{Structure parameter; has to be greater or equal than zero}

\item{rangex}{Vector with two components giving the interval of the real-line where linear part of the density should be evaluated.}

\item{T}{If plot=TRUE, parameter controlling the radius of the cylinder.}

\item{plot}{Logical; if TRUE, the 3D plot is produced.}

\item{axis}{Logical; if TRUE, axis are plotted.}

\item{orientation}{If plot=TRUE, character string indicating the orientation of the cylinder. Must be "horiz" or "vert".}
}
\value{
An list containing the following components: \item{fx}{ The estimated values of the density.}
\item{eval.points}{The points where the estimated density was evaluated.}
}
\description{
Function \code{dMS.cyl} computes the density function of a Mardia-Sutton density and produces an interactive 3D representation of the density on the surface of a cylinder.
}
\examples{
ms<-dMS.cyl(mu=c(pi,0),kappa=5,sigma=0.5,nu=1,lambda=3,rangex=c(-1,6),orientation="vert")
}
\references{
Mardia, K. V. & Sutton, T. W. (1978), ‘A model for cylindrical variables with applications’, J. Roy. Stat. Soc. Ser. B 40, 229–233.
}
