% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.density.models.R
\name{dsin.torus}
\alias{dsin.torus}
\title{dsin.torus}
\usage{
dsin.torus(mu, kappa, rho, plot = TRUE, axis = TRUE)
}
\arguments{
\item{mu}{Vector with two components representing the mean direction of each marginal model.}

\item{kappa}{Vector with two components representing the concentration of each marginal model.}

\item{rho}{Numeric element containing the correlation between the two components.}

\item{plot}{Logical; if TRUE, the 3D plot is produced.}

\item{axis}{Logical; if TRUE, axis are plotted.}
}
\value{
An list containing the following components: \item{fx}{ The estimated values of the density.}
\item{eval.points}{The points where the estimated density was evaluated.}
}
\description{
Function \code{dsin.torus} computes the density function of a bivariate sine model and produces an interactive 3D representation of the density on the surface of a cylinder.
}
\examples{
si<-dsin.torus(mu=c(-pi/16,pi/4), kappa=c(4,5), rho=3)
}
\references{
Singh, H., Hnizdo, V. & Demchuk, E. (2002), ‘Probabilistic model for two dependent circular variables’, Biometrika 89, 719–723.
}
