% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.density.models.R
\name{dsinsk.torus}
\alias{dsinsk.torus}
\title{dsinsk.torus}
\usage{
dsinsk.torus(mu, kappa, rho, lambda, model = NULL, plot = TRUE, axis = TRUE)
}
\arguments{
\item{mu}{Vector with two components representing the mean direction of each marginal model.}

\item{kappa}{Vector with two components representing the concentration of each marginal model.}

\item{rho}{Numeric element containing the correlation between the two components.}

\item{lambda}{Vectorwith two components representing the skewness coefficients.}

\item{model}{Character value indicating a predefined toroidal symmetric density.  Implemented examples include model="sine", model="cosine", model="bwc".}

\item{plot}{Logical; if TRUE, the 3D plot is produced.}

\item{axis}{Logical; if TRUE, axis are plotted.}
}
\value{
An list containing the following components: \item{fx}{ The estimated values of the density.}
\item{eval.points}{The points where the estimated density was evaluated.}
}
\description{
Function \code{dsinsk.torus} computes the density function of a sine-skewed toroidal distribution and produces an interactive 3D representation of the density on the surface of a cylinder.
}
\examples{
ssk<-dsinsk.torus(mu=c(-pi/2,pi/8),kappa=c(5,5),rho=3,lambda=c(-0.9,0.1),model="sine")
}
\references{
Jose Ameijeiras-Alonso, Christophe Ley, Sine-skewed toroidal distributions and their application in protein bioinformatics, Biostatistics, Volume 23, Issue 3, July 2022, Pages 685–704.
}
