% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.density.models.R
\name{dvMF.sph}
\alias{dvMF.sph}
\title{dvMF.sph}
\usage{
dvMF.sph(mu, kappa, plot = TRUE, axis = TRUE)
}
\arguments{
\item{mu}{Vector with three components giving the mean of the density in cartesian coordinates.}

\item{kappa}{Parameter controlling the concentration of the density.}

\item{plot}{Logical; if TRUE, the 3D plot is produced.}

\item{axis}{Logical; if TRUE, axis are plotted.}
}
\value{
An list containing the following components: \item{fx}{ The estimated values of the density.}
\item{eval.points}{The points where the estimated density was evaluated.}
}
\description{
Function \code{dvMF.sph} computes the density function of a von Mises-Fisher density of dimension 2 and produces an interactive 3D representation of the density on the surface of a sphere.
}
\examples{
vm<-dvMF.sph(c(1,0,0),5)
}
\references{
Mardia, K. V. & Jupp, P. E. (2000), Directional Statistics, Wiley, New York.
}
