% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.density.models.R
\name{dwc.torus}
\alias{dwc.torus}
\title{dwc.torus}
\usage{
dwc.torus(mu, kappa, rho, plot = TRUE, axis = TRUE)
}
\arguments{
\item{mu}{Vector with two components representing the mean direction of each marginal model.}

\item{kappa}{Vector with two components representing the concentration of each marginal model.}

\item{rho}{Numeric element containing the correlation between the two components.}

\item{plot}{Logical; if TRUE, the 3D plot is produced.}

\item{axis}{Logical; if TRUE, axis are plotted.}
}
\value{
An list containing the following components: \item{fx}{ The estimated values of the density.}
\item{eval.points}{The points where the estimated density was evaluated.}
}
\description{
Function \code{dcos.torus} computes the density function of a bivariate wrapped Cauchy and produces an interactive 3D representation of the density on the surface of a cylinder.
}
\examples{
wc<-dwc.torus(mu=c(pi/2,0), kappa=c(0.5,0.15), rho=0.7)
}
\references{
Kato, S. & Pewsey, A. (2015), ‘A Möbius transformation-induced distribution on the torus’, Biometrika 102, 359–370
}
