% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.density.plot.R
\name{sph.density.plot}
\alias{sph.density.plot}
\title{sph.density.plot}
\usage{
sph.density.plot(x, fx, axis = TRUE)
}
\arguments{
\item{x}{Matrix with three columns containing the points where the density is evaluated in cartesian coordinates.}

\item{fx}{Vector containing the values of the density. The length must coincide with the number of rows of x.}

\item{axis}{Logical; if TRUE, the axis are plotted.}
}
\description{
Function \code{sph.density.plot} creates an interactive 3D representation of a given spherical density.
}
\examples{
# Constructing a grid on the sphere
N<-300
ele_grid<-seq(0,pi,length=N)
ori<-list()
ori[[1]]<-0
ori[[N]]<-0
for (i in 2:(N-1)){
  ori[[i]]<- seq(0,2*pi,length=floor(2*N*sin(ele_grid[i])))
}
lonxi<-as.numeric(lapply(ori,length))
both<-cbind(unlist(ori),rep(ele_grid,times=lonxi))
x<-DirStats::to_sph(both[,1], both[,2]) # conversion of coordinates
fx<-Directional::iagd(x,mu=c(1,0,0))
sph.density.plot(x,fx)
}
