% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sph.plot.R
\name{sph.plot}
\alias{sph.plot}
\title{sph.plot}
\usage{
sph.plot(
  x,
  col.points = "red",
  pch = 16,
  size = 4,
  axis = TRUE,
  zoom = 0.8,
  windowRect = NULL,
  col.sph = "honeydew1"
)
}
\arguments{
\item{x}{Matrix containing the data in cartesian coordinates, where the number of columns must be 3 and the number of rows is the number of observations..}

\item{col.points}{Vector containing the mean direction of the von Mises-Fisher guide.}

\item{pch}{Type of points to be plotted.}

\item{size}{Numerical value containing the concentration of the von Mises-Fisher guide.}

\item{axis}{Logical; if TRUE, the axis are plotted.}

\item{zoom}{Parameter controlling the zoom in the plot.}

\item{windowRect}{aa.}

\item{col.sph}{Character string indicating the desired color of the sphere.}
}
\description{
Function \code{sph.plot} creates an interactive 3D representation of spherical data on a sphere.
}
\examples{
 library(movMF)
n<-200
mu<-matrix(c(0,0,1,0,0,-1),ncol=3,byrow=TRUE)
k<-c(7,2)
probs<-c(0.85,0.15)
x<-rmovMF(n,k*mu,alpha=probs)
sph.plot(x,2,4)
}
