% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.density.plot.R
\name{torus.density.plot}
\alias{torus.density.plot}
\title{torus.density.plot}
\usage{
torus.density.plot(x, fx, axis = TRUE)
}
\arguments{
\item{x}{Matrix with three columns containing the points where the density is evaluated in cartesian coordinates.}

\item{fx}{Vector containing the values of the density. The length must coincide with the number of rows of x.}

\item{axis}{Logical; if TRUE, the axis are plotted.}
}
\description{
Function \code{torus.density.plot} creates an interactive 3D representation of a given toroidal density.
}
\examples{
N<-600
u = seq(0, 2 * pi, length.out = N)
v = seq(0, 2 * pi, length.out = N)
ll<-expand.grid(u,v)
u2<-ll[,1]
v2<-ll[,2]
x<-cbind((1 + 0.5 * cos(v2)) * cos(u2), (1 + 0.5 * cos(v2)) * sin(u2), 0.5 * sin(v2))
fx<-dirdensity:::dbwc(u2,v2,mu1=pi,mu2=0,kappa1=0.5,kappa2=0.15,rho=0.5)
torus.density.plot(x, fx)
}
