% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torus.plot.R
\name{torus.plot}
\alias{torus.plot}
\title{torus.plot}
\usage{
torus.plot(
  x,
  col.points = "red",
  pch = 16,
  size = 4,
  axis = TRUE,
  zoom = 0.7,
  userMatrix = NULL,
  windowRect = NULL,
  col.torus = "honeydew1",
  labelsx = NULL
)
}
\arguments{
\item{x}{Matrix with two columns containing the data, where the number of rows is the number of observations. The first column contains the azimuthal ange, while the second column contains the polar angle (both in radians).}

\item{col.points}{Vector containing the mean direction of the von Mises-Fisher guide.}

\item{pch}{Type of points to be plotted.}

\item{size}{Numerical value containing the concentration of the von Mises-Fisher guide.}

\item{axis}{Logical; if TRUE, axis are plotted.}

\item{zoom}{Parameter controlling the zoom in the plot.}

\item{userMatrix}{A 4 by 4 matrix describing user actions to display the scene. See \code{par3d}.}

\item{windowRect}{A vector of four values indicating the left, top, right and bottom of the displayed window (in pixels). Applies to the whole device. See \code{par3d}.}

\item{col.torus}{Character string indicating the desired color of the torus.}

\item{labelsx}{Numeric vector containing the values to be displayed as labels for the azimuth component.}
}
\description{
Function \code{torus.plot} creates an interactive 3D representation of toroidal data on the surface of a torus.
}
\examples{
library(circular)
a1<-rvonmises(150,0,5)
a2<-rvonmises(150,0,15)
x<-cbind(a1,a2)
torus.plot(x)
}
